feat: add log replication to Raft example
Not tested yet, it might be a problem that the lastReplicated timestamp
cannot be implemented as in the Maelstrom documentation, due to time
not advancing while handling a message (i.e. each handler is
discrete). Not sure if this is a problem, or how difficult it would
be to fix.